requirements.txt
Django==4.2.13
djangorestframework==3.15.1
environs==11.0.0
psycopg2==2.9.9
pillow==10.3.0